Disable tst_QObject::connectSlotsVMIClass on MSVC
Taking the address of virtual member function within a class
with multiple inheritence cause internal compiler error.

Can be reproduced with this testcase:

```
  struct Base {
    virtual void slot1() { }
  };
  void x() { (void)&Base::slot1;  }

  struct SecondBase
  {};

  struct ObjectMultipleInherit : public Base, public SecondBase{
    void slot1() {}
  };

  void func() {
    constexpr auto xx = &ObjectMultipleInherit::slot1;
  }
```
Strangly, the crash happens when assigning the value to xx only if
the address of the function in tha base class is used before (in x())

Note: unrelated, but we can also get a crash with this simple
construct which is fortunately not used in Verdigris.
```
  struct XXX {
    virtual void func() { }
    static constexpr auto x = &XXX::func;
  };
```