Speedup compilation and save memory when there are many property
Since we need to evaluate ResolveNotifySignal N² times, we better optimize
it by avoiding any useless instantiation

Should solve the problem reported in https://github.com/woboq/verdigris/issues/28

In MSVC it is important not to create too many instance of ResolveNotifySignal
otherwise the compiler goes out of memory. By removing the Idx from the template
parameter, we avoid too many instentiation. This optimisation is also slightly
effective on Clang.
However, because of a GCC bug, it cannot be done for GCC.